add test method repair and add new file test
